+++ pkg/compiler/lib/compiler_api.dart
@@ -108,6 +108,20 @@ abstract class CompilerInput {
   /// is expected to hold a zero element at the last position. If this is not
   /// the case, the entire data structure is copied before scanning.
   Future<Input> readFromUri(Uri uri, {InputKind inputKind = InputKind.UTF8});
+
+  /// Register that [uri] should be an `InputKind.UTF8` input with the
+  /// given [source] as its zero-terminated list of contents.
+  ///
+  /// If [uri] was read prior to this call, this registration has no effect,
+  /// otherwise it is expected that a future [readFromUri] will return the
+  /// contents provided here.
+  ///
+  /// The main purpose of this API is to assist in error reporting when
+  /// compiling from kernel binary files. Binary files embed the contents
+  /// of source files that may not be available on disk. By using these
+  /// registered contents, dart2js will be able to provide accurate line/column
+  /// information on an error.
+  void registerUtf8ContentsForDiagnostics(Uri uri, List<int> source);
 }
 
 /// Output types used in `CompilerOutput.createOutputSink`.