+++ runtime/vm/os_fuchsia.cc
@@ -415,14 +415,6 @@ int OS::GetTimeZoneOffsetInSeconds(int64_t seconds_since_epoch) {
   return status == ZX_OK ? local_offset + dst_offset : 0;
 }
 
-int OS::GetLocalTimeZoneAdjustmentInSeconds() {
-  int32_t local_offset, dst_offset;
-  int64_t now_seconds = GetCurrentTimeNanos() / ZX_SEC(1);
-  zx_status_t status =
-      GetLocalAndDstOffsetInSeconds(now_seconds, &local_offset, &dst_offset);
-  return status == ZX_OK ? local_offset : 0;
-}
-
 int64_t OS::GetCurrentTimeMillis() {
   return GetCurrentTimeNanos() / ZX_MSEC(1);
 }